upd: move files to reflect duality api/cli on documentation
