+++ src/nflogic/api/cache.py
@@ -3,8 +3,8 @@ import logging
 from pathlib import Path
 import os
 
-from nflogic.parse import ParserInput, FactParser, FullParser, ParserInitError
-from nflogic import db
+from nflogic.api.parse import ParserInput, FactParser, FullParser, ParserInitError
+from nflogic.api import db
 
 SCRIPT_PATH = os.path.split(os.path.realpath(__file__))[0]
 CACHE_PATH = os.path.join(SCRIPT_PATH, "cache")
